+++ src/main/java/org/compiere/product/X_S_ResourceType.java
@@ -60,7 +60,7 @@ public class X_S_ResourceType extends BasePONameValue implements I_S_ResourceTyp
      * @return Tax Category
      */
     public int getTaxCategoryId() {
-        Integer ii = (Integer) getValue(COLUMNNAME_C_TaxCategory_ID);
+        Integer ii = getValue(COLUMNNAME_C_TaxCategory_ID);
         if (ii == null) return 0;
         return ii;
     }
@@ -71,7 +71,7 @@ public class X_S_ResourceType extends BasePONameValue implements I_S_ResourceTyp
      * @return Unit of Measure
      */
     public int getUOMId() {
-        Integer ii = (Integer) getValue(COLUMNNAME_C_UOM_ID);
+        Integer ii = getValue(COLUMNNAME_C_UOM_ID);
         if (ii == null) return 0;
         return ii;
     }
@@ -110,7 +110,7 @@ public class X_S_ResourceType extends BasePONameValue implements I_S_ResourceTyp
      * @return Category of a Product
      */
     public int getProductCategoryId() {
-        Integer ii = (Integer) getValue(COLUMNNAME_M_Product_Category_ID);
+        Integer ii = getValue(COLUMNNAME_M_Product_Category_ID);
         if (ii == null) return 0;
         return ii;
     }
@@ -219,7 +219,7 @@ public class X_S_ResourceType extends BasePONameValue implements I_S_ResourceTyp
      * @return Resource Type
      */
     public int getResourceTypeId() {
-        Integer ii = (Integer) getValue(COLUMNNAME_S_ResourceType_ID);
+        Integer ii = getValue(COLUMNNAME_S_ResourceType_ID);
         if (ii == null) return 0;
         return ii;
     }
